move parser under tools
